moved useful functions to process_manager
