+++ src/common/src/types/event/attributes/process.rs
@@ -28,7 +28,6 @@ pub struct FullProcessProperties {
     pub process_disk_usage_read_total: u64,
     pub process_disk_usage_write_total: u64,
     pub process_status: String,
-    pub input_files: Option<Vec<InputFile>>,
     pub container_id: Option<String>,
     pub job_id: Option<String>,
     pub working_directory: Option<String>,
